+++ domain/BarCodeManage.java
@@ -23,7 +23,7 @@ public class BarCodeManage implements Serializable {
     @Id
     @GeneratedValue
     @Column(nullable = false, name = "bar_code_id")
-    private Integer barCodeId;
+    private Long barCodeId;
     /**
      * 条形码值
      */
